+++ src/main/java/use_case/comment/CommentInteractor.java
@@ -1,7 +1,7 @@
 package use_case.comment;
 
 import data_access.DBUserDataAccessObject;
-import use_case.user_profile_user_story.note.DataAccessException;
+import use_case.DataAccessException;
 
 import java.util.List;
